Disable JAWTTest on Linux too.
From the source comment it seems that OSX and OpenJDK 1.6 are known to
be problematic, but I also get the failure on OpenJDK 1.7.
Assuming that it doesn't just work on Linux (or it could be working on
Oracle JDK but no idea how to detect and fix that).
So far disabling this.